+++ immutable/recurrentjava/flop/unary/SigmoidUnit.java
@@ -5,14 +5,14 @@ public class SigmoidUnit implements Unaflop {
 	private static final long serialVersionUID = 1L;
 
 	@Override
-	public double forward(double x) {
-		return 1 / (1 + Math.exp(-x));
+	public float forward(float x) {
+		return (float)(1 / (1 + Math.exp(-x)));
 	}
 
 	//benrayfield renamed this to deriv
 	@Override
-	public double deriv(double x) {
-		double act = forward(x);
+	public float deriv(float x) {
+		float act = forward(x);
 		return act * (1 - act);
 	}
